+++ src/trait_impl.rs
@@ -1,4 +1,10 @@
-//! This module contains the implementation of the Judge and Moral traits for common types
+/*! Implementation of the Judge and Moral traits for common types
+
+This module implements in order
+- Judge for Option, Result, ValRet and Moral
+- Return for impl Judge
+- (f="experimental") Try for ValRet and Moral
+*/
 use crate::*;
 
 // Implementation of Judge for Option, Result, ValRet and Moral
@@ -8,7 +14,7 @@ impl<T> Judge for Option<T> {
 	type Positive = T;
 	type Negative = ();
 
-	// #[inline]
+	#[inline]
 	fn into_moral (self) -> Moral<T, ()> {
 		match self {
 			Some(v) => Good(v),
@@ -16,11 +22,8 @@ impl<T> Judge for Option<T> {
 		}
 	}
 
-	#[inline]
-	fn from_good(v :T) -> Self { Some(v) }
-
-	#[inline]
-	fn from_bad(_ :()) -> Self { None }
+	#[inline] fn from_good(v :T) -> Self { Some(v) }
+	#[inline] fn from_bad(_ :()) -> Self { None }
 }
 
 /// Implementation of Judge for Result based on its implementation of Try
@@ -28,7 +31,7 @@ impl<T, E> Judge for Result<T, E> {
 	type Positive = T;
 	type Negative = E;
 
-	// #[inline]
+	#[inline]
 	fn into_moral (self) -> Moral<T, E> {
 		match self {
 			Ok(v) => Good(v),
@@ -36,17 +39,15 @@ impl<T, E> Judge for Result<T, E> {
 		}
 	}
 
-	#[inline]
-	fn from_good (v :T) -> Self { Ok(v) }
-
-	#[inline]
-	fn from_bad (v :E) -> Self { Err(v) }
+	#[inline] fn from_good (v :T) -> Self { Ok(v) }
+	#[inline] fn from_bad (v :E) -> Self { Err(v) }
 }
 
 impl<T, R> Judge for ValRet<T, R> {
 	type Positive = T;
 	type Negative = R;
 
+	#[inline]
 	fn into_moral (self) -> Moral<T, R> {
 		match self {
 			Val(v) => Good(v),
@@ -54,11 +55,8 @@ impl<T, R> Judge for ValRet<T, R> {
 		}
 	}
 
-	#[inline]
-	fn from_good (v :T) -> Self { Val(v) }
-
-	#[inline]
-	fn from_bad (r :R) -> Self { Ret(r) }
+	#[inline] fn from_good (v :T) -> Self { Val(v) }
+	#[inline] fn from_bad (r :R) -> Self { Ret(r) }
 }
 
 impl<Y, N> Judge for Moral<Y, N> {
@@ -68,11 +66,8 @@ impl<Y, N> Judge for Moral<Y, N> {
 	#[inline]
 	fn into_moral (self) -> Moral<Y, N> { self }
 
-	#[inline]
-	fn from_good (v :Y) -> Self { Good(v) }
-
-	#[inline]
-	fn from_bad (v :N) -> Self { Bad(v) }
+	#[inline] fn from_good (v :Y) -> Self { Good(v) }
+	#[inline] fn from_bad (v :N) -> Self { Bad(v) }
 }
 
 // Implementation of Return for those that implement Judge
@@ -84,4 +79,38 @@ impl<T, E, Me> Return for Me where Me :Judge<Positive = T, Negative = E> {
 	fn valret (self) -> ValRet<T, E> {
 		self.into_moral().into_valret()
 	}
+}
+
+// Implementation of Try for ValRet and Moral
+
+#[cfg(feature = "experimental")]
+impl<T, R> std::ops::Try for ValRet<T, R> {
+	type Ok = T;
+	type Error = R;
+
+	#[inline] fn into_result (self) -> Result<T, R> {
+		match self {
+			Val(v) => Ok(v),
+			Ret(r) => Err(r),
+		}
+	}
+
+	#[inline] fn from_ok (v :T) -> Self { Val(v) }
+	#[inline] fn from_error (v :R) -> Self { Ret(v) }
+}
+
+#[cfg(feature = "experimental")]
+impl<Y, N> std::ops::Try for Moral<Y, N> {
+	type Ok = Y;
+	type Error = N;
+
+	#[inline] fn into_result (self) -> Result<Y, N> {
+		match self {
+			Good(v) => Ok(v),
+			Bad(v) => Err(v),
+		}
+	}
+
+	#[inline] fn from_ok (v :Y) -> Self { Good(v) }
+	#[inline] fn from_error (v :N) -> Self { Bad(v) }
 }